DB 컬럼변경에 맞춰 타입 변경

+++ src/main/java/com/ocare/Ocare/domain/model/HealthRecordDetail.java
@@ -27,9 +27,9 @@ public class HealthRecordDetail {
     private LocalDateTime periodFrom; // PK 및 파티션 키
     private LocalDateTime periodTo;
     private String distanceUnit;
-    private Double distanceValue;
+    private BigDecimal distanceValue;
     private String caloriesUnit;
-    private Double caloriesValue;
+    private BigDecimal caloriesValue;
     private String createdId;
 
     @ManyToOne(fetch = FetchType.LAZY)
